+++ python/main.py
@@ -6,7 +6,7 @@ import webview
 
 from api import Api
 
-__version__ = '1.0.4'
+__version__ = '1.0.5'
 
 
 def get_public_path() -> Path: